Use pre-built locket binary and relative path for tps-watcher
Locket cannot be built when go modules are enabled. As a temporary
solution while the locket maintainers update to use modules, the tests
now expect the locket binary to be built independently and made
available on the $PATH.

Also, passing the code.cloudfoundry.org path when building tps-watcher doesn't
work with go mod and `gexec.Build`. This is a [known
limitation](https://github.com/onsi/gomega/issues/340) with Gomega.

Co-authored-by: Tom Viehman <tviehman@pivotal.io>
Co-authored-by: Weyman Fung <weymanf@vmware.com>